--- src/main/java/vista/jPanelEditing.java
@@ -1,154 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package vista;
-
-/**
- *
- * @author victo
- */
-public class jPanelEditing extends javax.swing.JFrame {
-
-    /**
-     * Creates new form jPanelEditing
-     */
-    public jPanelEditing() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        panelEdicion = new javax.swing.JPanel();
-        jLabeledicion = new javax.swing.JLabel();
-        jPanelPaneles = new javax.swing.JPanel();
-        jPanelEdicionBuscar = new javax.swing.JPanel();
-        edicionBuscar = new javax.swing.JButton();
-        jPanelEdicionReemplazar = new javax.swing.JPanel();
-        edicionReemplazar = new javax.swing.JButton();
-        jPanelEdicionSeleccionar = new javax.swing.JPanel();
-        edicionSeleccionar = new javax.swing.JButton();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        jLabeledicion.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabeledicion.setText("Edición");
-
-        javax.swing.GroupLayout panelEdicionLayout = new javax.swing.GroupLayout(panelEdicion);
-        panelEdicion.setLayout(panelEdicionLayout);
-        panelEdicionLayout.setHorizontalGroup(
-            panelEdicionLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jLabeledicion, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        panelEdicionLayout.setVerticalGroup(
-            panelEdicionLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelEdicionLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jLabeledicion, javax.swing.GroupLayout.DEFAULT_SIZE, 22, Short.MAX_VALUE))
-        );
-
-        jPanelPaneles.setLayout(new java.awt.GridLayout(3, 1));
-
-        jPanelEdicionBuscar.setLayout(new java.awt.BorderLayout());
-
-        edicionBuscar.setIcon(new javax.swing.ImageIcon("C:\\Users\\victo\\OneDrive\\Documentos\\NetBeansProjects\\iconosWordI\\iconosWordI\\Editar\\buscar (1).png")); // NOI18N
-        edicionBuscar.setText("Buscar");
-        edicionBuscar.setMargin(new java.awt.Insets(12, 14, 12, 14));
-        jPanelEdicionBuscar.add(edicionBuscar, java.awt.BorderLayout.CENTER);
-
-        jPanelPaneles.add(jPanelEdicionBuscar);
-
-        jPanelEdicionReemplazar.setLayout(new java.awt.BorderLayout());
-
-        edicionReemplazar.setIcon(new javax.swing.ImageIcon("C:\\Users\\victo\\OneDrive\\Documentos\\NetBeansProjects\\iconosWordI\\iconosWordI\\Editar\\reemplazarW.png")); // NOI18N
-        edicionReemplazar.setText("Reemplazar");
-        edicionReemplazar.setMargin(new java.awt.Insets(12, 14, 12, 14));
-        jPanelEdicionReemplazar.add(edicionReemplazar, java.awt.BorderLayout.CENTER);
-
-        jPanelPaneles.add(jPanelEdicionReemplazar);
-
-        jPanelEdicionSeleccionar.setLayout(new java.awt.BorderLayout());
-
-        edicionSeleccionar.setIcon(new javax.swing.ImageIcon("C:\\Users\\victo\\OneDrive\\Documentos\\NetBeansProjects\\iconosWordI\\iconosWordI\\Editar\\seleccionar (1).png")); // NOI18N
-        edicionSeleccionar.setText("Seleccionar");
-        edicionSeleccionar.setMargin(new java.awt.Insets(12, 14, 12, 14));
-        jPanelEdicionSeleccionar.add(edicionSeleccionar, java.awt.BorderLayout.CENTER);
-
-        jPanelPaneles.add(jPanelEdicionSeleccionar);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                    .addComponent(jPanelPaneles, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(panelEdicion, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jPanelPaneles, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(1, 1, 1)
-                .addComponent(panelEdicion, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(jPanelEditing.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(jPanelEditing.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(jPanelEditing.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(jPanelEditing.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new jPanelEditing().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton edicionBuscar;
-    private javax.swing.JButton edicionReemplazar;
-    private javax.swing.JButton edicionSeleccionar;
-    private javax.swing.JLabel jLabeledicion;
-    private javax.swing.JPanel jPanelEdicionBuscar;
-    private javax.swing.JPanel jPanelEdicionReemplazar;
-    private javax.swing.JPanel jPanelEdicionSeleccionar;
-    private javax.swing.JPanel jPanelPaneles;
-    private javax.swing.JPanel panelEdicion;
-    // End of variables declaration//GEN-END:variables
-}
