--- src/main/java/vista/jPanelStyles.java
@@ -1,512 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package vista;
-/**
- *
- * @author victo
- */
-public class jPanelStyles extends javax.swing.JFrame {
-
-    /**
-     * Creates new form jPanelStyles
-     */
-    public jPanelStyles() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        styles = new javax.swing.JLabel();
-        jPanelPaneles = new javax.swing.JPanel();
-        jPanelTittle = new javax.swing.JPanel();
-        lFuenteT = new javax.swing.JLabel();
-        lTitle = new javax.swing.JLabel();
-        jPanelSubtittle = new javax.swing.JPanel();
-        lFuenteSub = new javax.swing.JLabel();
-        lSubtitle = new javax.swing.JLabel();
-        jPanelSubtle = new javax.swing.JPanel();
-        lFuenteSubtle = new javax.swing.JLabel();
-        lSubtle = new javax.swing.JLabel();
-        jPanelEmphasis = new javax.swing.JPanel();
-        lFuenteEm = new javax.swing.JLabel();
-        lEmphasis = new javax.swing.JLabel();
-        jPanelIntense = new javax.swing.JPanel();
-        lFuenteIn = new javax.swing.JLabel();
-        lIntense = new javax.swing.JLabel();
-        jPanelTittle1 = new javax.swing.JPanel();
-        lFuenteT1 = new javax.swing.JLabel();
-        lTitle1 = new javax.swing.JLabel();
-        jPanelSubtittle1 = new javax.swing.JPanel();
-        lFuenteSub1 = new javax.swing.JLabel();
-        lSubtitle1 = new javax.swing.JLabel();
-        jPanelSubtle1 = new javax.swing.JPanel();
-        lFuenteSubtle1 = new javax.swing.JLabel();
-        lSubtle1 = new javax.swing.JLabel();
-        jPanelEmphasis1 = new javax.swing.JPanel();
-        lFuenteEm1 = new javax.swing.JLabel();
-        lEmphasis1 = new javax.swing.JLabel();
-        jPanelIntense1 = new javax.swing.JPanel();
-        lFuenteIn1 = new javax.swing.JLabel();
-        lIntense1 = new javax.swing.JLabel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setName("jFrameStyles"); // NOI18N
-
-        styles.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        styles.setText("Estilos");
-
-        jPanelTittle.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteT.setFont(new java.awt.Font("Segoe UI", 0, 48)); // NOI18N
-        lFuenteT.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteT.setText("AaB");
-
-        lTitle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lTitle.setText("Titulo");
-
-        javax.swing.GroupLayout jPanelTittleLayout = new javax.swing.GroupLayout(jPanelTittle);
-        jPanelTittle.setLayout(jPanelTittleLayout);
-        jPanelTittleLayout.setHorizontalGroup(
-            jPanelTittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelTittleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelTittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteT, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lTitle, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelTittleLayout.setVerticalGroup(
-            jPanelTittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelTittleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteT, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lTitle)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelSubtittle.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteSub.setFont(new java.awt.Font("Segoe UI", 0, 18)); // NOI18N
-        lFuenteSub.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteSub.setText("AaBbCcD");
-
-        lSubtitle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lSubtitle.setText("Subtitulo");
-
-        javax.swing.GroupLayout jPanelSubtittleLayout = new javax.swing.GroupLayout(jPanelSubtittle);
-        jPanelSubtittle.setLayout(jPanelSubtittleLayout);
-        jPanelSubtittleLayout.setHorizontalGroup(
-            jPanelSubtittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtittleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelSubtittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteSub, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lSubtitle, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelSubtittleLayout.setVerticalGroup(
-            jPanelSubtittleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtittleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteSub, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lSubtitle)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelSubtle.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteSubtle.setFont(new java.awt.Font("Segoe UI", 2, 18)); // NOI18N
-        lFuenteSubtle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteSubtle.setText("AaBbCcDd");
-
-        lSubtle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lSubtle.setText("Énfasis");
-
-        javax.swing.GroupLayout jPanelSubtleLayout = new javax.swing.GroupLayout(jPanelSubtle);
-        jPanelSubtle.setLayout(jPanelSubtleLayout);
-        jPanelSubtleLayout.setHorizontalGroup(
-            jPanelSubtleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelSubtleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteSubtle, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lSubtle, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelSubtleLayout.setVerticalGroup(
-            jPanelSubtleLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtleLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteSubtle, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lSubtle)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelEmphasis.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteEm.setFont(new java.awt.Font("Segoe UI", 3, 18)); // NOI18N
-        lFuenteEm.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteEm.setText("AaBbCcDd");
-
-        lEmphasis.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lEmphasis.setText("Énfasis intenso");
-
-        javax.swing.GroupLayout jPanelEmphasisLayout = new javax.swing.GroupLayout(jPanelEmphasis);
-        jPanelEmphasis.setLayout(jPanelEmphasisLayout);
-        jPanelEmphasisLayout.setHorizontalGroup(
-            jPanelEmphasisLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelEmphasisLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelEmphasisLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteEm, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lEmphasis, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelEmphasisLayout.setVerticalGroup(
-            jPanelEmphasisLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelEmphasisLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteEm, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lEmphasis)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelIntense.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteIn.setFont(new java.awt.Font("Segoe UI", 2, 18)); // NOI18N
-        lFuenteIn.setForeground(new java.awt.Color(51, 102, 255));
-        lFuenteIn.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteIn.setText("AaBbCcDd");
-
-        lIntense.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lIntense.setText("Énfasis comentario");
-
-        javax.swing.GroupLayout jPanelIntenseLayout = new javax.swing.GroupLayout(jPanelIntense);
-        jPanelIntense.setLayout(jPanelIntenseLayout);
-        jPanelIntenseLayout.setHorizontalGroup(
-            jPanelIntenseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelIntenseLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelIntenseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteIn, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lIntense, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelIntenseLayout.setVerticalGroup(
-            jPanelIntenseLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelIntenseLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteIn, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lIntense)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelTittle1.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteT1.setFont(new java.awt.Font("Segoe UI", 0, 18)); // NOI18N
-        lFuenteT1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteT1.setText("AaBbCcDd");
-
-        lTitle1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lTitle1.setText("Normal");
-
-        javax.swing.GroupLayout jPanelTittle1Layout = new javax.swing.GroupLayout(jPanelTittle1);
-        jPanelTittle1.setLayout(jPanelTittle1Layout);
-        jPanelTittle1Layout.setHorizontalGroup(
-            jPanelTittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelTittle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelTittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteT1, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lTitle1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelTittle1Layout.setVerticalGroup(
-            jPanelTittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelTittle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteT1, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lTitle1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelSubtittle1.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteSub1.setFont(new java.awt.Font("DejaVu Sans", 0, 24)); // NOI18N
-        lFuenteSub1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteSub1.setText("AaBbCcD");
-
-        lSubtitle1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lSubtitle1.setText("Título 1");
-
-        javax.swing.GroupLayout jPanelSubtittle1Layout = new javax.swing.GroupLayout(jPanelSubtittle1);
-        jPanelSubtittle1.setLayout(jPanelSubtittle1Layout);
-        jPanelSubtittle1Layout.setHorizontalGroup(
-            jPanelSubtittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtittle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelSubtittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteSub1, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lSubtitle1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelSubtittle1Layout.setVerticalGroup(
-            jPanelSubtittle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtittle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteSub1, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lSubtitle1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelSubtle1.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteSubtle1.setFont(new java.awt.Font("Sitka Text", 2, 18)); // NOI18N
-        lFuenteSubtle1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteSubtle1.setText("AaBbCcDd");
-
-        lSubtle1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lSubtle1.setText("Referencia sutil");
-
-        javax.swing.GroupLayout jPanelSubtle1Layout = new javax.swing.GroupLayout(jPanelSubtle1);
-        jPanelSubtle1.setLayout(jPanelSubtle1Layout);
-        jPanelSubtle1Layout.setHorizontalGroup(
-            jPanelSubtle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelSubtle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteSubtle1, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lSubtle1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelSubtle1Layout.setVerticalGroup(
-            jPanelSubtle1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelSubtle1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteSubtle1, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lSubtle1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelEmphasis1.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteEm1.setFont(new java.awt.Font("Times New Roman", 1, 18)); // NOI18N
-        lFuenteEm1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteEm1.setText("AaBbCcDd");
-
-        lEmphasis1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lEmphasis1.setText("Título del libro");
-
-        javax.swing.GroupLayout jPanelEmphasis1Layout = new javax.swing.GroupLayout(jPanelEmphasis1);
-        jPanelEmphasis1.setLayout(jPanelEmphasis1Layout);
-        jPanelEmphasis1Layout.setHorizontalGroup(
-            jPanelEmphasis1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelEmphasis1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelEmphasis1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteEm1, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lEmphasis1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelEmphasis1Layout.setVerticalGroup(
-            jPanelEmphasis1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelEmphasis1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteEm1, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lEmphasis1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        jPanelIntense1.setPreferredSize(new java.awt.Dimension(150, 100));
-
-        lFuenteIn1.setFont(new java.awt.Font("Segoe UI", 0, 18)); // NOI18N
-        lFuenteIn1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lFuenteIn1.setText("AaBbCcDd");
-
-        lIntense1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        lIntense1.setText("Párrafo de lista");
-
-        javax.swing.GroupLayout jPanelIntense1Layout = new javax.swing.GroupLayout(jPanelIntense1);
-        jPanelIntense1.setLayout(jPanelIntense1Layout);
-        jPanelIntense1Layout.setHorizontalGroup(
-            jPanelIntense1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelIntense1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelIntense1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(lFuenteIn1, javax.swing.GroupLayout.DEFAULT_SIZE, 138, Short.MAX_VALUE)
-                    .addComponent(lIntense1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-        jPanelIntense1Layout.setVerticalGroup(
-            jPanelIntense1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelIntense1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(lFuenteIn1, javax.swing.GroupLayout.PREFERRED_SIZE, 70, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(lIntense1)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout jPanelPanelesLayout = new javax.swing.GroupLayout(jPanelPaneles);
-        jPanelPaneles.setLayout(jPanelPanelesLayout);
-        jPanelPanelesLayout.setHorizontalGroup(
-            jPanelPanelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelPanelesLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelPanelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanelPanelesLayout.createSequentialGroup()
-                        .addComponent(jPanelTittle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelSubtittle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelSubtle, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelEmphasis, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelIntense, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(jPanelPanelesLayout.createSequentialGroup()
-                        .addComponent(jPanelTittle1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelSubtittle1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelSubtle1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelEmphasis1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(jPanelIntense1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap())
-        );
-        jPanelPanelesLayout.setVerticalGroup(
-            jPanelPanelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanelPanelesLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanelPanelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanelTittle, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelSubtittle, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelSubtle, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelEmphasis, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelIntense, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(jPanelPanelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanelTittle1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelSubtittle1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelSubtle1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelEmphasis1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelIntense1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(20, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(styles, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGap(0, 13, Short.MAX_VALUE)
-                        .addComponent(jPanelPaneles, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jPanelPaneles, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(styles, javax.swing.GroupLayout.PREFERRED_SIZE, 38, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(jPanelStyles.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(jPanelStyles.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(jPanelStyles.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(jPanelStyles.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new jPanelStyles().setVisible(true);
-            }
-        });
-        
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JPanel jPanelEmphasis;
-    private javax.swing.JPanel jPanelEmphasis1;
-    private javax.swing.JPanel jPanelIntense;
-    private javax.swing.JPanel jPanelIntense1;
-    private javax.swing.JPanel jPanelPaneles;
-    private javax.swing.JPanel jPanelSubtittle;
-    private javax.swing.JPanel jPanelSubtittle1;
-    private javax.swing.JPanel jPanelSubtle;
-    private javax.swing.JPanel jPanelSubtle1;
-    private javax.swing.JPanel jPanelTittle;
-    private javax.swing.JPanel jPanelTittle1;
-    private javax.swing.JLabel lEmphasis;
-    private javax.swing.JLabel lEmphasis1;
-    private javax.swing.JLabel lFuenteEm;
-    private javax.swing.JLabel lFuenteEm1;
-    private javax.swing.JLabel lFuenteIn;
-    private javax.swing.JLabel lFuenteIn1;
-    private javax.swing.JLabel lFuenteSub;
-    private javax.swing.JLabel lFuenteSub1;
-    private javax.swing.JLabel lFuenteSubtle;
-    private javax.swing.JLabel lFuenteSubtle1;
-    private javax.swing.JLabel lFuenteT;
-    private javax.swing.JLabel lFuenteT1;
-    private javax.swing.JLabel lIntense;
-    private javax.swing.JLabel lIntense1;
-    private javax.swing.JLabel lSubtitle;
-    private javax.swing.JLabel lSubtitle1;
-    private javax.swing.JLabel lSubtle;
-    private javax.swing.JLabel lSubtle1;
-    private javax.swing.JLabel lTitle;
-    private javax.swing.JLabel lTitle1;
-    private javax.swing.JLabel styles;
-    // End of variables declaration//GEN-END:variables
-}
