--- src/main/java/vista/jPanelSuperior.java
@@ -1,317 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package vista;
-
-/**
- *
- * @author victo
- */
-public class jPanelSuperior extends javax.swing.JFrame {
-
-    /**
-     * Creates new form jPanelSuperior
-     */
-    public jPanelSuperior() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jTabbedPane = new javax.swing.JTabbedPane();
-        jtpFile = new javax.swing.JPanel();
-        jtpHome = new javax.swing.JPanel();
-        jPanelEstilos = new javax.swing.JPanel();
-        jPanelEditar = new javax.swing.JPanel();
-        jPanelPortapapeles = new javax.swing.JPanel();
-        jPanelFuente = new javax.swing.JPanel();
-        jPanelPárrafo = new javax.swing.JPanel();
-        jtpInsert = new javax.swing.JPanel();
-        jtpDesign = new javax.swing.JPanel();
-        jtpLayout = new javax.swing.JPanel();
-        jtpReferences = new javax.swing.JPanel();
-        jtpMailings = new javax.swing.JPanel();
-        jPanelScrollTop = new javax.swing.JPanel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        javax.swing.GroupLayout jtpFileLayout = new javax.swing.GroupLayout(jtpFile);
-        jtpFile.setLayout(jtpFileLayout);
-        jtpFileLayout.setHorizontalGroup(
-            jtpFileLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpFileLayout.setVerticalGroup(
-            jtpFileLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Archivo", jtpFile);
-
-        jPanelEstilos.setBackground(new java.awt.Color(102, 255, 204));
-
-        javax.swing.GroupLayout jPanelEstilosLayout = new javax.swing.GroupLayout(jPanelEstilos);
-        jPanelEstilos.setLayout(jPanelEstilosLayout);
-        jPanelEstilosLayout.setHorizontalGroup(
-            jPanelEstilosLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 199, Short.MAX_VALUE)
-        );
-        jPanelEstilosLayout.setVerticalGroup(
-            jPanelEstilosLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 187, Short.MAX_VALUE)
-        );
-
-        jPanelEditar.setBackground(new java.awt.Color(204, 255, 0));
-
-        javax.swing.GroupLayout jPanelEditarLayout = new javax.swing.GroupLayout(jPanelEditar);
-        jPanelEditar.setLayout(jPanelEditarLayout);
-        jPanelEditarLayout.setHorizontalGroup(
-            jPanelEditarLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 199, Short.MAX_VALUE)
-        );
-        jPanelEditarLayout.setVerticalGroup(
-            jPanelEditarLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 187, Short.MAX_VALUE)
-        );
-
-        jPanelPortapapeles.setBackground(new java.awt.Color(204, 255, 0));
-
-        javax.swing.GroupLayout jPanelPortapapelesLayout = new javax.swing.GroupLayout(jPanelPortapapeles);
-        jPanelPortapapeles.setLayout(jPanelPortapapelesLayout);
-        jPanelPortapapelesLayout.setHorizontalGroup(
-            jPanelPortapapelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 199, Short.MAX_VALUE)
-        );
-        jPanelPortapapelesLayout.setVerticalGroup(
-            jPanelPortapapelesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 187, Short.MAX_VALUE)
-        );
-
-        jPanelFuente.setBackground(new java.awt.Color(102, 255, 204));
-
-        javax.swing.GroupLayout jPanelFuenteLayout = new javax.swing.GroupLayout(jPanelFuente);
-        jPanelFuente.setLayout(jPanelFuenteLayout);
-        jPanelFuenteLayout.setHorizontalGroup(
-            jPanelFuenteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 199, Short.MAX_VALUE)
-        );
-        jPanelFuenteLayout.setVerticalGroup(
-            jPanelFuenteLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 187, Short.MAX_VALUE)
-        );
-
-        jPanelPárrafo.setBackground(new java.awt.Color(204, 255, 0));
-
-        javax.swing.GroupLayout jPanelPárrafoLayout = new javax.swing.GroupLayout(jPanelPárrafo);
-        jPanelPárrafo.setLayout(jPanelPárrafoLayout);
-        jPanelPárrafoLayout.setHorizontalGroup(
-            jPanelPárrafoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 199, Short.MAX_VALUE)
-        );
-        jPanelPárrafoLayout.setVerticalGroup(
-            jPanelPárrafoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 187, Short.MAX_VALUE)
-        );
-
-        javax.swing.GroupLayout jtpHomeLayout = new javax.swing.GroupLayout(jtpHome);
-        jtpHome.setLayout(jtpHomeLayout);
-        jtpHomeLayout.setHorizontalGroup(
-            jtpHomeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jtpHomeLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jPanelPortapapeles, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanelFuente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanelPárrafo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanelEstilos, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPanelEditar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        jtpHomeLayout.setVerticalGroup(
-            jtpHomeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jtpHomeLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jtpHomeLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanelPortapapeles, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelEditar, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelPárrafo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelFuente, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jPanelEstilos, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(61, Short.MAX_VALUE))
-        );
-
-        jTabbedPane.addTab("Inicio", jtpHome);
-
-        javax.swing.GroupLayout jtpInsertLayout = new javax.swing.GroupLayout(jtpInsert);
-        jtpInsert.setLayout(jtpInsertLayout);
-        jtpInsertLayout.setHorizontalGroup(
-            jtpInsertLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpInsertLayout.setVerticalGroup(
-            jtpInsertLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Insertar", jtpInsert);
-
-        javax.swing.GroupLayout jtpDesignLayout = new javax.swing.GroupLayout(jtpDesign);
-        jtpDesign.setLayout(jtpDesignLayout);
-        jtpDesignLayout.setHorizontalGroup(
-            jtpDesignLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpDesignLayout.setVerticalGroup(
-            jtpDesignLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Presentación", jtpDesign);
-
-        javax.swing.GroupLayout jtpLayoutLayout = new javax.swing.GroupLayout(jtpLayout);
-        jtpLayout.setLayout(jtpLayoutLayout);
-        jtpLayoutLayout.setHorizontalGroup(
-            jtpLayoutLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpLayoutLayout.setVerticalGroup(
-            jtpLayoutLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Revisar", jtpLayout);
-
-        javax.swing.GroupLayout jtpReferencesLayout = new javax.swing.GroupLayout(jtpReferences);
-        jtpReferences.setLayout(jtpReferencesLayout);
-        jtpReferencesLayout.setHorizontalGroup(
-            jtpReferencesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpReferencesLayout.setVerticalGroup(
-            jtpReferencesLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Vista", jtpReferences);
-
-        javax.swing.GroupLayout jtpMailingsLayout = new javax.swing.GroupLayout(jtpMailings);
-        jtpMailings.setLayout(jtpMailingsLayout);
-        jtpMailingsLayout.setHorizontalGroup(
-            jtpMailingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jtpMailingsLayout.setVerticalGroup(
-            jtpMailingsLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 254, Short.MAX_VALUE)
-        );
-
-        jTabbedPane.addTab("Ayuda", jtpMailings);
-
-        jPanelScrollTop.setBackground(new java.awt.Color(255, 0, 204));
-
-        javax.swing.GroupLayout jPanelScrollTopLayout = new javax.swing.GroupLayout(jPanelScrollTop);
-        jPanelScrollTop.setLayout(jPanelScrollTopLayout);
-        jPanelScrollTopLayout.setHorizontalGroup(
-            jPanelScrollTopLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 1030, Short.MAX_VALUE)
-        );
-        jPanelScrollTopLayout.setVerticalGroup(
-            jPanelScrollTopLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 36, Short.MAX_VALUE)
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(jPanelScrollTop, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addContainerGap())
-            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(layout.createSequentialGroup()
-                    .addContainerGap()
-                    .addComponent(jTabbedPane, javax.swing.GroupLayout.PREFERRED_SIZE, 1030, Short.MAX_VALUE)
-                    .addContainerGap()))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addContainerGap(297, Short.MAX_VALUE)
-                .addComponent(jPanelScrollTop, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(layout.createSequentialGroup()
-                    .addContainerGap()
-                    .addComponent(jTabbedPane, javax.swing.GroupLayout.PREFERRED_SIZE, 287, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addContainerGap(46, Short.MAX_VALUE)))
-        );
-
-        jTabbedPane.getAccessibleContext().setAccessibleName("");
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(jPanelSuperior.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(jPanelSuperior.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(jPanelSuperior.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(jPanelSuperior.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new jPanelSuperior().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JPanel jPanelEditar;
-    private javax.swing.JPanel jPanelEstilos;
-    private javax.swing.JPanel jPanelFuente;
-    private javax.swing.JPanel jPanelPortapapeles;
-    private javax.swing.JPanel jPanelPárrafo;
-    private javax.swing.JPanel jPanelScrollTop;
-    private javax.swing.JTabbedPane jTabbedPane;
-    private javax.swing.JPanel jtpDesign;
-    private javax.swing.JPanel jtpFile;
-    private javax.swing.JPanel jtpHome;
-    private javax.swing.JPanel jtpInsert;
-    private javax.swing.JPanel jtpLayout;
-    private javax.swing.JPanel jtpMailings;
-    private javax.swing.JPanel jtpReferences;
-    // End of variables declaration//GEN-END:variables
-}
